+++ server/webportal/src/main/java/ru/kolaer/server/webportal/mvc/model/dao/impl/TicketDaoImpl.java
@@ -19,7 +19,7 @@ public class TicketDaoImpl extends AbstractDefaultDao<TicketEntity> implements T
     }
 
     @Override
-    public List<TicketEntity> findAllByRegisterId(Integer id) {
+    public List<TicketEntity> findAllByRegisterId(Long id) {
         return getSession()
                 .createQuery("FROM " + getEntityName() + " t WHERE t.ticketRegister.id = :id", getEntityClass())
                 .setParameter("id", id)
